First refactoring of the notebook into separate files
